+++ src/component.js
@@ -1,0 +1,32 @@
+import {createElement} from './create-element.js';
+
+export default class Component {
+  constructor() {
+    if (new.target === Component) {
+      throw new Error(`Can't instantiate BaseComponent, only concrete one.`);
+    }
+
+    this._element = null;
+  }
+
+  get element() {
+    return this._element;
+  }
+
+  get template() {
+    throw new Error(`You have to define template.`);
+  }
+
+  render() {
+    if (!this._element) {
+      this._element = createElement(this.template);
+      this.bind();
+    }
+    return this._element;
+  }
+
+  bind() {}
+
+  unbind() {}
+
+}
